DeleteWaterModal added
