+++ package.json
@@ -10,11 +10,13 @@
     "@testing-library/user-event": "^13.5.0",
     "assert": "^2.1.0",
     "axios": "^1.7.9",
+    "bootstrap": "^5.3.3",
     "browserify-zlib": "^0.2.0",
     "caniuse-lite": "^1.0.30001688",
     "formik": "^2.4.6",
     "https-browserify": "^1.0.0",
     "react": "^18.1.0",
+    "react-bootstrap": "^2.10.7",
     "react-dom": "^18.1.0",
     "react-hot-toast": "^2.4.1",
     "react-modal": "^3.16.1",
